add info to the statuslines
