+++ src/helper/index.ts
@@ -22,6 +22,5 @@ export * from "./env";
 export * from "./fiatHelpers";
 export * from "./lamdas";
 export * from "./path";
-export * from "./printInfo";
 export * from "./reg-conversion";
 export * from "./toposort";
