+++ src/optimizer/optimizer.helper.ts
@@ -14,13 +14,13 @@
  * limitations under the License.
  */
 
-import { cpus } from "os";
+import os, { cpus } from "os";
 
 import { CHOICE } from "@/enums";
-import { bl, cy, env, gn, pu, rd, re, SI, yl } from "@/helper";
+import { bl, cy, env, gn, pu, rd, re, shouldProof, SI, yl } from "@/helper";
 import globals from "@/helper/globals";
 import { Model } from "@/model";
-import type { AnalyseResult } from "@/types";
+import type { AnalyseResult, OptimizerArgs } from "@/types";
 
 const { CC, CFLAGS } = env;
 
@@ -93,6 +93,8 @@ export function genStatistics(a: {
   numRevert: { [k: string]: number };
   numMut: { [k: string]: number };
   counter: string;
+  framePointer: string;
+  memoryConstraints: string;
 }): string[] {
   return [
     `; cpu ${cpus()[0].model}`,
@@ -100,6 +102,8 @@ export function genStatistics(a: {
     `; seed ${a.paddedSeed} `,
     `; CC / CFLAGS ${CC} / ${CFLAGS} `,
     `; using counter; ${a.counter}`,
+    `; framePointer ${a.framePointer}`,
+    `; memoryConstraints ${a.memoryConstraints}`,
     `; time needed: ${a.elapsed} ms on ${a.evals} evaluations.`,
     `; Time spent for assembling and measuring (initial batch_size=${a.batchSize}, initial num_batches=${a.numBatches}): ${a.acc} ms`,
     `; number of used evaluations: ${a.evals}`,
@@ -123,4 +127,43 @@ export function logMutation({
   const dDetails = choice == "D " ? Model.decisionStats : "             ";
 
   globals.mutationLog.push([numEvals, choice.trim(), kept ? 1 : 0, pDetails, dDetails].join(","));
+}
+
+export function printStartInfo({
+  resultDir,
+  bridge,
+  seed,
+  evals,
+  cyclegoal,
+  proof,
+  symbolname,
+  counter,
+  framePointer,
+  memoryConstraints,
+}: Pick<
+  OptimizerArgs,
+  "resultDir" | "bridge" | "seed" | "evals" | "cyclegoal" | "proof" | "framePointer" | "memoryConstraints"
+> & {
+  symbolname: string;
+  counter: string;
+}) {
+  process.stdout.write(
+    [
+      `\nStart`,
+      `on brg/symbolname>>${cy}${bridge}/${symbolname}${re}<<`,
+      `>>${cy}${shouldProof({ bridge, proof }) ? "with" : "without"} proofing${re} correct<<`,
+      `on cpu >>${cy}${os.cpus()[0].model}${re}<<`,
+      `writing results to>>${cy}${resultDir}${re}<<`,
+      `with seed >>${cy}${seed}${re}<<`,
+      `for >>${cy}${SI(evals)}${re}<< evaluations`,
+      `against CC>>${cy}${CC} ${CFLAGS}${re}<<`,
+      `with cycle goal>>${cy}${cyclegoal}${re}<< for each measurement`,
+      `on host>>${cy}${os.hostname()}${re}<<`,
+      `with pid>>${cy}${process.pid}${re}<<`,
+      `using counter>>${cy}${counter}${re}<<`,
+      `framePointer=>>${cy}${framePointer}${re}<<`,
+      `memoryConstraints>>${cy}${memoryConstraints}${re}<<`,
+      `starting @>>${cy}${new Date().toISOString()}${re}<<\n`,
+    ].join(" "),
+  );
 }
